+++ nv/vi/text_objects.py
@@ -27,7 +27,11 @@ from sublime import CLASS_WORD_START
 from sublime import IGNORECASE
 from sublime import Region
 
+from NeoVintageous.nv.polyfill import view_find
+from NeoVintageous.nv.polyfill import view_indentation_level
+from NeoVintageous.nv.polyfill import view_indented_region
 from NeoVintageous.nv.utils import next_non_blank
+from NeoVintageous.nv.utils import next_non_ws
 from NeoVintageous.nv.utils import prev_non_blank
 from NeoVintageous.nv.utils import prev_non_ws
 from NeoVintageous.nv.vi.search import find_in_range
@@ -60,7 +64,8 @@ WORD = 5
 BIG_WORD = 6
 PARAGRAPH = 7
 INDENT = 8
-LINE = 9
+BIG_INDENT = 9
+LINE = 10
 
 
 PAIRS = {
@@ -82,6 +87,7 @@ PAIRS = {
     't': (None, TAG),
     'W': (None, BIG_WORD),
     'w': (None, WORD),
+    'I': (None, BIG_INDENT),
     'i': (None, INDENT),
     'l': (None, LINE),
 }  # type: dict
@@ -388,11 +394,9 @@ def get_text_object_region(view, s, text_object, inclusive=False, count=1):
             return Region(sentence_start, sentence_end.b)
 
     # Support for a port of the Indent Object plugin:
-    # https://github.com/michaeljsmith/vim-indent-object
-    if type_ == INDENT:
-        start, end = find_indent_text_object(view, s, inclusive)
-
-        return Region(start, end)
+    # A port of https://github.com/michaeljsmith/vim-indent-object.
+    if type_ in (INDENT, BIG_INDENT):
+        resolve_indent_text_object(view, s, inclusive, big=(type_ == BIG_INDENT))
 
     if type_ == LINE:
         start, end = find_line_text_object(view, s)
@@ -635,101 +639,71 @@ def find_inner_paragraph(view, initial_loc):
     return (begin, end)
 
 
-def find_indent_text_object(view, s, inclusive=True):
-    """Implement the indent text object as specified at http://vim.wikia.com/wiki/Indent_text_object."""
-    begin = view.line(s)
-    end = view.line(s)
+def resolve_indent_text_object(view, s, inclusive=True, big=False):
 
-    start_line = view.line(s)
-    start_line_content = view.substr(start_line)
+    # Look for the minimum indentation in the current visual region.
+    idnt = 1000
+    idnt_pt = None
+    for line in view.lines(s):
+        if not re.match('^\\s*$', view.substr(line)):
+            level = view.indentation_level(line.a)
+            if level < idnt:
+                idnt = min(idnt, level)
+                idnt_pt = line.a
 
-    # Do nothing when the line is whitespace-only
-    if re.match("^\\s*$", start_line_content):
-        return (s.a, s.b)
+    # If the selection has no indentation at all, find which indentation level
+    # is the largest, the previous non blank before tphe cursor or the next non
+    # blank after the cursor, and start the selection from that point.
 
-    whitespace_match = re.match("\\s*", start_line_content)
-    if not whitespace_match:
-        return (s.a, s.b)
+    if idnt == 1000:
+        pnb_pt = prev_non_ws(view, s.begin())
+        pnb_indent_level = view_indentation_level(view, pnb_pt)
 
-    whitespace = whitespace_match.group(0)
-    whitespace_length = len(whitespace)
+        nnb_pt = next_non_ws(view, s.end())
+        nnb_indent_level = view_indentation_level(view, nnb_pt)
 
-    # From http://vim.wikia.com/wiki/Indent_text_object:
-    if inclusive:
-        # "a"
-        if whitespace_length == 0:
-            # When the cursor is on a line with zero indent,
-            # the selection will be delimited by blank lines
-            # (that may or may not contain whitespaces).
-            pattern = None
-            break_on_empty_lines = True
+        if pnb_indent_level > nnb_indent_level:
+            idnt_pt = s.a = s.b = pnb_pt
+        elif nnb_indent_level > pnb_indent_level:
+            idnt_pt = s.a = s.b = nnb_pt
         else:
-            # When the cursor is on a line with a non-zero indent,
-            # the selection will be delimited by lines with
-            # an indent that is less than the original line;
-            # blank lines will be SELECTED.
-            pattern = "\\s{0," + str(whitespace_length - 1) + "}\\S"
-            break_on_empty_lines = False
-    else:
-        # "i"
-        if whitespace_length == 0:
-            # When the cursor is on a line with zero indent,
-            # the selection will be delimited by blank lines
-            # (that may or may not contain whitespaces).
-            pattern = None
-            break_on_empty_lines = True
-        else:
-            # When the cursor is on a line with a non-zero indent,
-            # the selection will be delimited by lines with
-            # an indent that is less than the original line;
-            # blank lines will be IGNORED and thus,
-            # become the delimiter if one is encountered.
-            pattern = "\\s{0," + str(whitespace_length - 1) + "}\\S"
-            break_on_empty_lines = True
+            idnt_pt = pnb_pt
 
-    if pattern:
-        compiled_pattern = re.compile(pattern)
+    if idnt == 0 and idnt_pt is not None:
+        expanded = view.expand_by_class(s, CLASS_EMPTY_LINE)
+        s.a = expanded.a
+        s.b = expanded.b
 
-        def should_break_on_line(line_content):
-            if break_on_empty_lines and not line_content.strip():
-                return True
+        if not inclusive:
+            # Case: ii and iI. Strip any leading whitespace.
+            leading_ws = view_find(view, '\\s*', s.a)
+            if leading_ws is not None:
+                s.a = view.line(leading_ws.b).a
 
-            return "pattern_match" if compiled_pattern.match(line_content) else False
-    else:
-        def should_break_on_line(line_content):
-            return not line_content.strip()
+            s.b = prev_non_blank(view, s.b)
+        elif big:
+            # Case: aI. Add a line below.
+            if view.substr(s.b) == '\n':
+                s.b += 1
 
-    # Search backward until the pattern is matched
-    p = s.a
-    while True:
-        line = view.line(p)
-        line_content = view.substr(line)
-        if should_break_on_line(line_content):
-            break
-        elif line.begin() == 0:
-            p = 0
-            break
+    elif idnt > 0 and idnt_pt is not None:
+        indented_region = view_indented_region(view, idnt_pt, inclusive)
 
-        p = line.begin() - 1
+        if indented_region.begin() < s.begin():
+            s.a = indented_region.begin()
 
-    begin = p + 1 if p > 0 else p
+        if indented_region.end() > s.end():
+            s.b = indented_region.end()
 
-    # To get the value for end, we do the same thing, this time searching forward.
-    p = s.b
-    while True:
-        line = view.line(p)
-        line_content = view.substr(line)
-        if should_break_on_line(line_content):
-            break
+        if inclusive:
+            # Case: ai. Add a line above.
+            s.a = view.line(view.text_point(view.rowcol(s.a)[0] - 1, 0)).a
 
-        p = line.end() + 1
+            # Case: aI. Add a line below.
+            if big:
+                s.b = view.full_line(view.text_point(view.rowcol(s.b - 1)[0] + 1, 0)).b
 
-        if p >= view.size():
-            break
-
-    end = p - 1
-
-    return (begin, end)
+    return s
 
 
 def find_line_text_object(view, s):